Amazingly I got to a place where I can use this. Thus, huge rewrite!
Using a lot of my new knowledge in ruby/rails to improve on the original
code. Simplified a lot of the gem related code to the bare minimum to cut
down on the noise of everything going on. Still need to write up specs to
unit test the library, but for now... No tests :O