Added options; Better scope; Post create and update callbacks
Can configure to return default value when approval is missing. This is
also the value used as the default for the approval when it is first
created.

Also specify a value to set the approval after the record is updated.

Wrapped scope in lambda so that SQL calls don't go out during init.

Set the approval association to destroy approvals if the record is removed.